+++ crates/router/src/connector/bluesnap/transformers.rs
@@ -17,11 +17,7 @@ use crate::{
     consts,
     core::errors,
     pii::Secret,
-    types::{
-        self, api,
-        storage::enums,
-        transformers::{ForeignInto, ForeignTryFrom},
-    },
+    types::{self, api, storage::enums, transformers::ForeignTryFrom},
     utils::{Encode, OptionExt},
 };
 
@@ -353,7 +349,7 @@ impl TryFrom<types::PaymentsSessionResponseRouterData<BluesnapWalletTokenRespons
                             ),
                         payment_request_data: Some(api_models::payments::ApplePayPaymentRequest {
                             country_code: item.data.get_billing_country()?,
-                            currency_code: item.data.request.currency.foreign_into(),
+                            currency_code: item.data.request.currency,
                             total: api_models::payments::AmountInfo {
                                 label: applepay_metadata.data.payment_request_data.label,
                                 total_type: Some("final".to_string()),