+++ crates/router/src/core/admin.rs
@@ -25,7 +25,6 @@ use crate::{
             types::{self as domain_types, AsyncLift},
         },
         storage,
-        transformers::ForeignInto,
     },
     utils::{self, OptionExt},
 };
@@ -503,7 +502,7 @@ pub async fn create_payment_connector(
 
     let merchant_connector_account = domain::MerchantConnectorAccount {
         merchant_id: merchant_id.to_string(),
-        connector_type: req.connector_type.foreign_into(),
+        connector_type: req.connector_type,
         connector_name: req.connector_name.to_string(),
         merchant_connector_id: utils::generate_id(consts::ID_LENGTH, "mca"),
         connector_account_details: domain_types::encrypt(
@@ -670,7 +669,7 @@ pub async fn update_payment_connector(
 
     let payment_connector = storage::MerchantConnectorAccountUpdate::Update {
         merchant_id: None,
-        connector_type: Some(req.connector_type.foreign_into()),
+        connector_type: Some(req.connector_type),
         connector_name: None,
         merchant_connector_id: None,
         connector_account_details: req