+++ crates/router/src/core/payments/flows/authorize_flow.rs
@@ -86,7 +86,7 @@ impl Feature<api::Authorize, types::PaymentsAuthorizeData> for types::PaymentsAu
                 resp.to_owned(),
                 maybe_customer,
                 merchant_account,
-                self.request.payment_method_type.clone(),
+                self.request.payment_method_type,
             )
             .await?;