+++ crates/router/src/core/payments/operations/payment_confirm.rs
@@ -102,7 +102,6 @@ impl<F: Send + Clone> GetTracker<F, PaymentData<F>, api::PaymentsRequest> for Pa
 
         payment_intent.setup_future_usage = request
             .setup_future_usage
-            .map(ForeignInto::foreign_into)
             .or(payment_intent.setup_future_usage);
 
         let (token, payment_method, payment_method_type, setup_mandate, mandate_connector) =
@@ -145,13 +144,9 @@ impl<F: Send + Clone> GetTracker<F, PaymentData<F>, api::PaymentsRequest> for Pa
 
         payment_attempt.payment_experience = request
             .payment_experience
-            .map(|experience| experience.foreign_into())
             .or(payment_attempt.payment_experience);
 
-        payment_attempt.capture_method = request
-            .capture_method
-            .or(payment_attempt.capture_method.map(|cm| cm.foreign_into()))
-            .map(|cm| cm.foreign_into());
+        payment_attempt.capture_method = request.capture_method.or(payment_attempt.capture_method);
 
         currency = payment_attempt.currency.get_required_value("currency")?;
         amount = payment_attempt.amount.into();
@@ -391,8 +386,8 @@ impl<F: Clone> UpdateTracker<F, PaymentData<F>, api::PaymentsRequest> for Paymen
             .straight_through_algorithm
             .clone();
         let payment_token = payment_data.token.clone();
-        let payment_method_type = payment_data.payment_attempt.payment_method_type.clone();
-        let payment_experience = payment_data.payment_attempt.payment_experience.clone();
+        let payment_method_type = payment_data.payment_attempt.payment_method_type;
+        let payment_experience = payment_data.payment_attempt.payment_experience;
         let additional_pm_data = payment_data
             .payment_method_data
             .as_ref()